+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### PaymentSheet
 * [Fixed][8710](https://github.com/stripe/stripe-android/pull/8710) Fixed issue where no payment method was selected after navigating back to the select saved payment method screen.
+* [ADDED][8717](https://github.com/stripe/stripe-android/pull/8717) Add CVC Recollection functionality to `PaymentSheet` and `PaymentSheet.FlowController`
 
 ## 20.47.4 - 2024-06-27